requirements.txt
langchain-groq
python-dotenv
pywhatkit